requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.3
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
Django==3.0.7
django-appconf==1.0.4
django-model-utils==4.0.0
django-notifications-hq==1.6.0
idna==2.9
isort==4.3.21
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.6.0
pylint==2.5.2
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
stripe==2.48.0
swapper==1.1.2.post1
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
